+++ lib/vite.config.ts
@@ -3,7 +3,7 @@ import { resolve } from 'path';
 import { defineConfig } from 'vite';
 import dts from 'vite-plugin-dts';
 import { visualizer } from "rollup-plugin-visualizer";
-import { existsSync, readdirSync, lstatSync, rmdirSync, unlinkSync, writeFileSync, copyFileSync } from 'fs'
+import { existsSync, readdirSync, lstatSync, rmdirSync, unlinkSync, writeFileSync, copyFileSync, readFileSync } from 'fs'
 import libCss from 'vite-plugin-libcss';
 
 const generateComponentsEntries = () => {
@@ -33,7 +33,95 @@ function emptyDir(dir: string): void {
 }
 
 emptyDir(resolve(__dirname, 'dist'))
-emptyDir(resolve(__dirname, 'types'))
+
+const beforeWriteFile = (path: string, content: string) => {
+    //Replace by the name of the folder
+    const splittedPath = path.split('/').map(p => p);
+    const lastItem = splittedPath[splittedPath.length - 1];
+    const beforeLastItem = splittedPath[splittedPath.length - 2];
+    const component = splittedPath[splittedPath.length - 3];
+    if (lastItem === 'index.d.ts') {
+        //Write an copy file, but at one upper level with the name of the beforeLastItem
+        const createContent = (path: string) => `export * from './dist/${path}/index'`;
+        console.log(splittedPath.join('/'));
+        if (beforeLastItem === 'dist' || beforeLastItem === 'styles') return;
+        let newContent = component.toLowerCase() === 'components' ? createContent(`components/${beforeLastItem}`) : createContent(beforeLastItem);
+        writeFileSync(`${__dirname}/${beforeLastItem}.d.ts`, newContent);
+    }
+    return { path, content };
+}
+
+const removeWrongCSSImport = (file: string) => {
+    const filesToIncludeCSS = ['Tooltip', 'StickerCard', 'LifecycleProgress'];
+    const fileNameWithoutType = file.replace('.cjs.js', '').replace('.es.js', '');
+    console.log(fileNameWithoutType);
+    const isFileToIncludeCSS = filesToIncludeCSS.includes(fileNameWithoutType)
+    if(!isFileToIncludeCSS) {
+        //Remove the `import './styles.css'` from the file.
+        const filePath = resolve(__dirname, 'dist', file);
+        let content = readFileSync(filePath, 'utf8');
+        content = content.replace("import './style.css';", '');
+        content = content.replace("require('./style.css');", '');
+        writeFileSync(filePath, content, 'utf8');
+        console.warn("Removed CSS import from file: ", file);
+    }
+    else console.warn("File: ", file, " is not removed from CSS import");
+}
+
+const purgeReactDom = () => {
+    //Access every dist file that ends with .js and remove every `import "react-dom" or require("react-dom");`
+    const distFiles = readdirSync(resolve(__dirname, 'dist'));
+    const jsFiles = distFiles.filter(f => f.endsWith('.js'));
+    for (const file of jsFiles) {
+        const filePath = resolve(__dirname, 'dist', file);
+        let content = readFileSync(filePath, 'utf8');
+        content = content.replace(/import "react-dom";/g, '');
+        content = content.replace(/require\("react-dom"\);/g, '');
+        content = content.replace(/import "react";/g, '');
+        content = content.replace(/require\("react"\);/g, '');
+        content = content.replace(/import "react\/dom";/g, '');
+        content = content.replace(/require\("react\/dom"\);/g, '');
+        content = content.replace(/import "react\/index";/g, '');
+        content = content.replace(/require\("react\/index"\);/g, '');
+        writeFileSync(filePath, content, 'utf8');
+        removeWrongCSSImport(file);
+    }
+}
+
+const generateDeclarationTypes = () => {
+    //Copy every `[file].d.ts` from dist to the root of __dirname and then create a index.d.ts that ///references this files
+    const files = readdirSync(resolve(__dirname));
+    const dtsFiles = files.filter(f => f.endsWith('.d.ts'));
+    const packageJson = require(resolve(__dirname, 'package.json'));
+    packageJson.files = ['dist'];
+    packageJson.exports = {};
+    for (const file of dtsFiles) {
+        const destPath = resolve(__dirname, `${file}`);
+        copyFileSync(resolve(__dirname, file), destPath);
+        //Then update package.json `files` to include all this files
+        packageJson.files = [...packageJson.files, file];
+        //update all package.json `exports` to include the keys with the file name with import and require
+        const exports = packageJson.exports;
+        const exportKeyName = file.startsWith('index') ? '.' : "./" + file.replace('.d.ts', '');
+        exports[exportKeyName] = {
+            import: `./dist/${file.replace('.d.ts', '.es.js')}`,
+            require: `./dist/${file.replace('.d.ts', '.cjs.js')}`
+        }
+        writeFileSync(resolve(__dirname, 'package.json'), JSON.stringify(packageJson, null, 2));
+    }
+    let indexContent = '';
+    for (const file of dtsFiles) {
+        if (file !== 'index.d.ts')
+            indexContent += `/// <reference path="${file}" />\n export * from './${file.replace('.d.ts', '')}'\n`; // 
+    }
+    writeFileSync(resolve(__dirname, 'index.d.ts'), indexContent);
+}
+
+
+const afterBuild = () => {
+    purgeReactDom();
+    generateDeclarationTypes();
+}
 
 export default defineConfig({
     plugins: [
@@ -47,52 +135,8 @@ export default defineConfig({
             //insertTypesEntry: true,
             skipDiagnostics: true,
             logDiagnostics: true,
-            beforeWriteFile: (path, content) => {
-                //console.log(path);
-                //Replace by the name of the folder
-                const splittedPath = path.split('/').map(p => p);
-                const lastItem = splittedPath[splittedPath.length - 1];
-                const beforeLastItem = splittedPath[splittedPath.length - 2];
-                const component = splittedPath[splittedPath.length - 3];
-                if (lastItem === 'index.d.ts') {
-                    //Write an copy file, but at one upper level with the name of the beforeLastItem
-                    const createContent = (path: string) => `export * from './dist/${path}/index'`;
-                    console.log(splittedPath.join('/'));
-                    if(beforeLastItem === 'dist' || beforeLastItem === 'styles') return;
-                    let newContent = component.toLowerCase() === 'components' ? createContent(`components/${beforeLastItem}`) : createContent(beforeLastItem);
-                    writeFileSync(`${__dirname}/${beforeLastItem}.d.ts`, newContent);
-                }
-                return { path, content };
-            },
-            afterBuild: () => {
-                //Copy every `[file].d.ts` from dist to the root of __dirname and then create a index.d.ts that ///references this files
-                const files = readdirSync(resolve(__dirname));
-                const dtsFiles = files.filter(f => f.endsWith('.d.ts'));
-                const packageJson = require(resolve(__dirname, 'package.json'));
-                packageJson.files = ['dist'];
-                packageJson.exports = {};
-                for (const file of dtsFiles) {
-                    const destPath = resolve(__dirname, `${file}`);
-                    copyFileSync(resolve(__dirname, file), destPath);
-                    //Then update package.json `files` to include all this files
-                    packageJson.files = [...packageJson.files, file];
-                    //update all package.json `exports` to include the keys with the file name with import and require
-                    const exports = packageJson.exports;
-                    const exportKeyName = file.startsWith('index') ? '.' : "./" + file.replace('.d.ts', '');
-                    exports[exportKeyName] = {
-                        import: `./dist/${file.replace('.d.ts', '.es.js')}`,
-                        require: `./dist/${file.replace('.d.ts', '.cjs.js')}`
-                    }
-                    writeFileSync(resolve(__dirname, 'package.json'), JSON.stringify(packageJson, null, 2));
-                }
-                let indexContent = '';
-                for (const file of dtsFiles) {
-                    if (file !== 'index.d.ts')
-                        indexContent += `/// <reference path="${file}" />\n export * from './${file.replace('.d.ts', '')}'\n`; // 
-                }
-                writeFileSync(resolve(__dirname, 'index.d.ts'), indexContent);
-
-            }
+            beforeWriteFile,
+            afterBuild
         }),
         visualizer({
             gzipSize: true,